Using short for nbLoc

+++ webservice-rest/src/main/java/etu/polytech/ws/rest/data/Activite.java
@@ -14,7 +14,7 @@ public class Activite  {
     private ActiviteId id;
 
     @Column(name = "nbloc")
-    private int nbLoc;
+    private short nbLoc;
 
     public Activite() {}
